Fix gift upload handling with multer middleware
