+++ server/routes/giftRoutes.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const router = express.Router();
+const upload = require('../utils/upload');
 const { createGift, getGift } = require('../controllers/giftController');
 
 const asyncRoute = (handler) => (req, res, next) => {
@@ -9,7 +10,7 @@ const asyncRoute = (handler) => (req, res, next) => {
   });
 };
 
-router.post('/', asyncRoute(createGift));
+router.post('/', upload.single('video'), asyncRoute(createGift));
 router.get('/:id', asyncRoute(getGift));
 
 module.exports = router;
